fix clang-tidy

+++ tasks/all/kholin_k_multidimensional_integrals_rectangle/src/ops_all.cpp
@@ -17,7 +17,7 @@ double kholin_k_multidimensional_integrals_rectangle_all::TestTaskALL::Integrate
 
   double sum = 0.0;
   for (size_t i = 0; i < static_cast<size_t>(n); ++i) {
-    f_values_[curr_index_dim] = l_limits[curr_index_dim] + (i + 0.5) * h[curr_index_dim];
+    f_values_[curr_index_dim] = l_limits[curr_index_dim] + (static_cast<double>(i) + 0.5) * h[curr_index_dim];
     sum += Integrate(f_, l_limits, u_limits, h, f_values_, curr_index_dim + 1, dim_, n);
   }
   return sum * h[curr_index_dim];
